backend/tools_deployment/tb-cloudrun/requirements.txt
flask
gunicorn
torch
opencv-python-headless
pillow
numpy
